+++ framework/py/flwr/superlink/routers/control/router_test.py
@@ -18,8 +18,8 @@ from fastapi import FastAPI
 from fastapi.testclient import TestClient
 
 from flwr.proto.control_pb2 import ListRunsRequest  # pylint: disable=E0611
+from flwr.supercore.routers.protobuf import PROTOBUF_MEDIA_TYPE
 from flwr.superlink.routers.control.router import router
-from flwr.superlink.routers.router import PROTOBUF_MEDIA_TYPE
 
 
 def test_list_runs_is_a_protobuf_endpoint() -> None:
